moved products selections to common components
