+++ src/components/index.ts
@@ -5,3 +5,4 @@ export { default as Searchbar } from './SearchBar/SearchBar';
 export { default as data } from './SearchBar/data';
 export { default as Location } from './Location/Location';
 export { default as Payment } from './Payment/Payment';
+export { default as Product } from './Product/Product';
