+++ src/pages/PostJobs/PostJobs.tsx
@@ -1,21 +1,26 @@
 import React from 'react';
-import { ErrorBoundary } from '../../components';
-import { useLocation } from 'react-router-dom';
+import { ErrorBoundary, Product, Payment } from '../../components';
+import { useLocation, useHistory } from 'react-router-dom';
+import  ProductData  from '../../components/DataModels/ProductData';
 import Form from './Form';
 interface Props{
-  product:{
-    type:string,
-    amount:number,
-    hostingTime:number,
-    offers:{field:string,value:boolean}[]
-  }
+  product:ProductData
 }
 const PostJobs = () => {
-  let location= useLocation();
-  console.log(location.state);
+  const location= useLocation();
+  const history =  useHistory();
+  const product = location.state;
+  if(!product){
+    history.push('/PostJobs');
+  }
   return (
     <ErrorBoundary>
-      <Form/>
+      <div className="job_page">
+        <div className="content">
+          <Product product={ product } isSelected={true} />
+          <Payment />
+        </div>
+      </div>
     </ErrorBoundary>
   );
 };
