+++ src/pages/ProductSelectionPage/ProductSelectionPage.scss
@@ -15,90 +15,3 @@
     border: 0px solid #000;
     border-radius: 5px;
 }
-.content .col{
-    border: 1px solid #000;
-    padding: 10px;
-    min-width: 32%;
-    border: 1px solid rgba(255, 255, 255, 1);
-    border-radius: 5px;
-    text-align: center;
-    justify-content: space-around;
-    border: 1px solid lightgray;
-    padding-bottom: 30px;
-    background-color: #fff;
-    transition: .5s;
-    margin-top: 20px;
-}
-.content .col:hover{
-    border: 1px solid lightgrey;
-    box-shadow: 0 6px 16px 0 rgba(0, 0, 0, 0.1);
-    background-color: rgb(229,235,254);
-    transform: translateY(-20px);
-}
-.content .col .price{   
-    align-items: center;
-    justify-content: space-around;
-    margin-left: auto;
-    margin-right: auto;
-    max-width: 80%;
-    border-bottom: 2px solid lightgray;
-}
-.content .col h3{
-    margin: 10px 0px;
-    padding: 5px 0px;
-    border: 0px solid #000;
-    font-weight: 600;
-    font-size: 35px;
-    color: #fff;
-    letter-spacing: 1px;
-    border-radius: 0px;
-    color: rgb(254,172,41);
-}
-.content .col h3 small{
-    font-size: 15px;
-    font-weight: 400;
-}
-.content .col .heading{
-    text-align: center;
-    padding: 0px 20px;
-    color: rgb(31, 4, 110);
-    font-weight: 600;
-    font-size: 40px;
-    margin-left: auto;
-    margin-right: auto;
-    border-radius: 0px;
-}
-.content .col p{
-    padding: 10px 30px;
-    text-align: left;
-}
-.content .col p span{
-    font-weight: 700;
-    font-size: 17px;
-    padding-left: 0;
-    float: right;
-}
-.content .col p .yes{
-    color: rgb(6, 100, 6);
-}
-.content .col p .no{
-    color: rgb(255, 0, 0);
-}
-.content .col .btn a{
-    text-decoration: none;
-    font-weight: 400;
-    font-size: 15px;
-    padding: 8px 50px;
-    border-radius: 5px;
-    background-color: rgb(80,37,209);   
-    color: #fff;
-}
-.content .col .btn:hover{
-    transform: scale(1.1);
-}
-.content .col .btn{
-    padding-top: 10px;
-    transition: 0.5s;
-    cursor: pointer;
-}
-
